add simple error handling for product and update type and cart product controller

+++ controllers/cartProductController.js
@@ -7,6 +7,7 @@ class CartProductController {
       const cartProducts = await CartProductService.getAll(cartId);
       res.json(cartProducts);
     } catch (err) {
+      console.error({ err });
       res.status(500).json(err);
     }
   }
@@ -16,6 +17,7 @@ class CartProductController {
       const cartProduct = await CartProductService.getOne(id, cartId);
       res.json(cartProduct);
     } catch (err) {
+      console.log({ err });
       res.status(500).json(err);
     }
   }
@@ -27,15 +29,23 @@ class CartProductController {
       const cartProduct = await CartProductService.create(cartId, productId);
       res.json(cartProduct);
     } catch (err) {
+      console.error({ err });
       res.status(500).json(err.message);
     }
   }
   async deleteCartProduct(req, res) {
     try {
       const { id, cartId } = req.params;
-      const cartProduct = await CartProductService.delete(id, cartId);
-      res.json(cartProduct);
+      const candidate = await CartProductService.getOne(id, cartId);
+      if (!candidate) {
+        return res
+          .status(400)
+          .json({ message: 'Cart product with this id doesnt exist' });
+      }
+      await CartProductService.delete(id, cartId);
+      res.json({ message: 'Cart product was deleted' });
     } catch (err) {
+      console.error({ err });
       res.status(500).json(err.message);
     }
   }
@@ -45,7 +55,9 @@ class CartProductController {
       const { productId } = req.body;
       const candidate = await CartProductService.getOne(id, cartId);
       if (!candidate) {
-        res.status(400).json({ message: 'No cart product with this id' });
+        res
+          .status(400)
+          .json({ message: 'Cart product with this id doesnt exist' });
       }
       const cartProduct = await CartProductService.update(
         id,
@@ -54,6 +66,7 @@ class CartProductController {
       );
       res.json(cartProduct);
     } catch (err) {
+      console.error({ err });
       res.status(500).json(err.message);
     }
   }
